+++ src/components/EditSections/EditUserInfo/constants.js
@@ -1,0 +1,9 @@
+export const USER_TYPES = {
+  staff: 'staff',
+  patron: 'patron',
+  shadow: 'shadow',
+};
+
+export const checkIfConsortiumEnabled = stripes => {
+  return stripes.hasInterface('consortia');
+};